+++ source/extensions/lightspeed.trex.logic.ogn/python/nodes/InterpolateFloat.py
@@ -1,0 +1,14 @@
+# GENERATED FILE - DO NOT EDIT
+# This file is a stub for OmniGraph editor compatibility, and is not used by the Remix Runtime.
+from __future__ import annotations
+
+from typing import TYPE_CHECKING
+
+if TYPE_CHECKING:
+    from lightspeed.trex.logic.ogn.ogn.InterpolateFloatDatabase import InterpolateFloatDatabase
+
+
+class InterpolateFloat:
+    @staticmethod
+    def compute(_db: InterpolateFloatDatabase):
+        return True