+++ source/extensions/lightspeed.trex.app.resources/docs/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.22.1]
+### Changed
+- Updated default layout configurations with graph editor window
+
 ## [1.22.0]
 ### Added
 - Added new sidebar icons for workspace navigation (AI Tools, Home, Ingestion, Modding, Package Mod, Project Setup)